Rather archive Builds on Tree
The build object details still contains the origin commit.
We still know where the build has been triggered but the more important
is to know if the tree builds. (The commit message/author/date doesn't
breaks the build.)

By the way, all the build result persistence is now done in BuildDAOImpl.