Program options as positional options are now allowed.
Nevertheless, there's still one case to implement that it's not yet
considered. When using the magrit executable (magrit), if
an unrecognized options is passed *before* the positional parameters,
it will be accepted. E.g.:

  magrit build -1 log

will be considered the same as:

  magrit build log -1

We are not using the magrit executable, but just in case.