Getting prepared to compile for other platforms.
By getting stricter with compilation warnings, some porting bugs
can be early detected (e.g. typical bug is to use an integer
that gets truncated in win32 but not in linux64. It can be difficult
to spot once the code is written.  -Wconversion avoids that).

I had to fix very few lines of boost-process to stick to the new
stricter rules.